fix ddpg data type error

+++ ddpg.py
@@ -87,7 +87,8 @@ def train(mu, mu_target, q, q_target, memory, q_optimizer, mu_optimizer):
     s,a,r,s_prime,done_mask  = memory.sample(batch_size)
     
     target = r + gamma * q_target(s_prime, mu_target(s_prime))
-    q_loss = F.smooth_l1_loss(q(s,a), target.detach())
+    
+    q_loss = F.smooth_l1_loss(q(s,a.float()), target.detach().float())
     q_optimizer.zero_grad()
     q_loss.backward()
     q_optimizer.step()
